removed apikey in favor of a feature flag to enable the internal routes
