+++ packages/core/lib/config/config.ts
@@ -9,6 +9,10 @@ export interface Configuration {
   EXT_API_BASE_URL: string;
   EXT_API_MOCKING: boolean;
 
+  // INTERNAL TEST API
+  INTERNAL_ROUTES_ENABLED: boolean;
+  INTERNAL_ROUTES_TEST_USER: string[];
+
   // PDND Interoperability
   INTEROP_API_MOCKING: boolean;
   INTEROP_AUTH_SERVER_ENDPOINT_URL: string;
@@ -20,22 +24,23 @@ export interface Configuration {
   INTEROP_CLIENT_ASSERTION_SUB: string;
   INTEROP_CLIENT_ASSERTION_TYPE: string;
   INTEROP_CLIENT_ID: string;
+
   INTEROP_ESERVICE_AUDIENCE: string;
   INTEROP_GRANT_TYPE: string;
-
   IS_BROWSER: boolean;
   IS_DEVELOPMENT: boolean;
   IS_MSW_ENABLED: boolean; // Mock Service Worker - API mocking library for browser and Node.js
   IS_PRODUCTION: boolean;
+
   IS_SERVER: boolean;
   IS_TEST: boolean;
-
   // FIMS OIDC
   OIDC_CLIENT_ID: string;
   OIDC_CLIENT_REDIRECT_URI: string;
   OIDC_CLIENT_SECRET: string;
   OIDC_ISSUER_URL: string;
   OIDC_MOCK_FORCED_ENABLE?: boolean;
+
   OIDC_MOCK_JWT_PRIVATE_KEY?: string;
   OIDC_MOCK_JWT_PUBLIC_KEY?: string;
 }
@@ -48,6 +53,13 @@ export const getConfiguration = (): Configuration => ({
   EXT_API_BASE_PATH: process.env.EXT_API_BASE_PATH as string,
   EXT_API_BASE_URL: process.env.EXT_API_BASE_URL as string,
   EXT_API_MOCKING: process.env.EXT_API_MOCKING === "true",
+
+  INTERNAL_ROUTES_ENABLED: process.env.INTERNAL_ROUTES_ENABLED === "true",
+  INTERNAL_ROUTES_TEST_USER: process.env.INTERNAL_ROUTES_TEST_USER
+    ? process.env.INTERNAL_ROUTES_TEST_USER.split(",").map((user) =>
+        user.trim(),
+      )
+    : [],
 
   INTEROP_API_MOCKING: process.env.INTEROP_API_MOCKING === "true",
   INTEROP_AUTH_SERVER_ENDPOINT_URL: process.env
