+++ packages/core/lib/wrappers/index.ts
@@ -1,5 +1,5 @@
-export * from "./with-api-key-and-voucher-handler";
-export * from "./with-api-key-handler";
+export * from "./internal/with-test-user-and-voucher-internal-handler";
+export * from "./internal/with-test-user-internal-handler";
 export * from "./with-jwt-auth-handler";
 export * from "./with-jwt-auth-voucher-handler";
 export * from "./with-voucher-handler";
